+++ Source/Mosa.Compiler.Framework/Stages/PlatformEdgeSplitStage.cs
@@ -6,7 +6,7 @@ namespace Mosa.Compiler.Framework.Stages;
 ///	This stage removes critical edges by inserting empty basic blocks. Some SSA optimizations and the flow
 ///	control resolution in the register allocator require that all critical edges are removed.
 /// </summary>
-public class PlatformEdgeSplitStage : BaseEdgeSplitStage
+public class PlatformEdgeSplitStage : EdgeSplitStage
 {
 	/// <summary>
 	/// Inserts the jump instruction.
